@@ -9,8 +9,8 @@ DATA = PluginData('robot')
 
 
 class Config(BaseSettings):
-    tencent_ai_app_id: str = DATA.config.get('tencent', 'app_id')
-    tencent_ai_app_key: str = DATA.config.get('tencent', 'app_key')
+    tencent_ai_secret_id: str = DATA.config.get('tencent', 'secret_id')
+    tencent_ai_secret_key: str = DATA.config.get('tencent', 'secret_key')
     tuling_api_key: str = DATA.config.get('tuling', 'api_key')
 
     class Config:
